// frontend/src/app/api-client/model/order.ts
/**
 * Order API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Order { 
    readonly id?: number;
    petId?: number;
    quantity?: number;
    shipDate?: string;
    orderStatus?: Order.OrderStatusEnum;
    complete?: boolean;
}
export namespace Order {
    export type OrderStatusEnum = 'PLACED' | 'APPROVED' | 'DELIVERED';
    export const OrderStatusEnum = {
        Placed: 'PLACED' as OrderStatusEnum,
        Approved: 'APPROVED' as OrderStatusEnum,
        Delivered: 'DELIVERED' as OrderStatusEnum
    };
}